fix: any-state transitions and handle events strictly one at a time
This change fixes a couple of bugs

* any-state transitions were previously only working when the machine's current state was
  mentioned in the machine definition. This change allows any-state transitions to work
  regardless of whether the current state is mentioned in the machine definition.
* event-handling was previously incorrectly re-entrant, meaning that it was possible
  to start handling a new event while already handling an event
  (eg, an `onTransition` side-effect could dispatch an event to the machine while it was
  handling the event that triggered the transition). Now we ensure that events are handled
  strictly one at a time.

Also added an "elevators" example that uses an autonomous standalone machine for each elevator,
plus a controller machine which composes and manages multiple elevator machines. This is the most
complex example so far, and demonstrates how to use machines to model a system with multiple
sub-systems that interact with each other. (And it helped write tests for the bugs above.)

One day I'd like to add a visualisation of this as I think it would be a good teaching tool,
not least showing how to integrate machines with a UI.