fix: types enforce the `enableCopyDataOnTransition` machine-definition option vs `data()` callback
When a state-machine's state-data is homogenous (ie, for all state, every state-type property
expect `name` is the same), we offer the additional `enableCopyDataOnTransition?: boolean`
machine-definition-config option.

> ```typescript
>  /**
>   * If `true`, data is automatically copied between states when transitioning, and the
>   * transition does not provide its own `data()` callback implementation.
>   * This avoids boilerplate `data()` callbacks config, like `{ to: 'foo', data: ({ state }) => state }`.
>   * @default false
>   */
>  readonly enableCopyDataOnTransition: true;
> ```

Previously it was possible to define a machine *without setting `enableCopyDataOnTransition: true`*,
**but also not provide `data()` callbacks**, which is clearly unsafe from a type-safety perspective,
but also resulted in buggy or unexpected behaviour at run-time.

This change to the types means that it should now be enforced that the (homogenous-state) machine is
EITHER defined with `enableCopyDataOnTransition: true` and MAY to provide `data()` callbacks,
OR defined with `enableCopyDataOnTransition: false | undefined` and MUST provide `data()` callbacks.