docs: sync examples in docs from example source files
This change adds tooling to automate keeping the examples in the docs (markdown files) up-to-date.

`scripts/sync-docs.ts` now searches the docs for "💡 View this example..." (links to GitHub) then
extracts the example's filename and updates the source code in the docs from the file.

It should also fail the build when the docs are out-of-date vs example sources,
so examples in READMEs etc should be working from now on.