fix: handle side-effects sending events to the machine while starting and stopping (#5)
This change aims to make machines more resilient to edge-cases like using side-effects
to send events to the machine while it is starting or stopping.

When starting any events sent by side-effects should be queued and handled
after the machine is started.

When stopping any events sent by side-effects should be ignored.