fix: ignore events from expired side-effects
Previously it was possible for a side-effect to remain "alive" after it's expected lifetime
and continue to send the machine events (with the captured `send()` function from the
side-effect function parameter).

This shouldn't be possible, and only active side-effects should be able to send events
to machine.

The `send()` function received by side-effects is now proxied and any events passed to
this function after the side-effect is expired (cleaned-up, disposed) are ignored.