docs: link directly to GitHub pages from `yay-machine` package README.md
This README.md is published with the package and welcomes people on `npmjs.com`
so I think the best place to send them is now the GitHub pages versions of the docs.

They are already better (IMHO) than the vanilla GitHub pages experience, and
they can be made even better over time.